requirements.txt
pytrends>=5.0.0
pandas>=1.0.0
requests>=2.25.0
urllib3==1.26.18
lxml>=4.6.0
matplotlib>=3.3.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.8.0
python-dotenv>=1.0.0
pytz>=2023.3
wbgapi>=1.0.0
sdmx>=1.0.0

